source_code/requirements.txt
mlflow==3.9.0
databricks-agents==1.9.3
databricks-langchain==0.14.0
databricks-vectorsearch==0.64
langgraph==1.0.6
langchain==1.2.6
databricks-sdk==0.82.0
databricks-sdk[openai]==0.82.0
pandas
pyyaml==6.0.3
cloudpickle==3.1.2
flask==3.1.2
setuptools==80.10.2
databricks-openai
databricks-cli
openai